refactor: move add desktop layout into
