+++ src/client/pages/Layout/MobileLayout.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+
+export const MobileLayout: React.FC = React.memo(() => {
+  return <div>Mobile</div>;
+});
+MobileLayout.displayName = 'MobileLayout';
