+++ src/client/pages/Layout/types.ts
@@ -1,0 +1,3 @@
+export interface LayoutProps extends React.PropsWithChildren {
+  list?: React.ReactNode;
+}
